+++ VKDesign(11September)/Models/Models for InformationItem/Contacts.swift
@@ -1,0 +1,15 @@
+//
+//  Contacts.swift
+//  VKDesign(11September)
+//
+//  Created by BLVCK on 25/09/2017.
+//  Copyright © 2017 blvvvck production. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+struct Contacts {
+    var image: UIImage
+    var contact: String
+}
