+++ VKDesign(11September)/Models/Models for InformationItem/EducationalEstablishments.swift
@@ -1,0 +1,14 @@
+//
+//  EducationalEstablishments.swift
+//  VKDesign(11September)
+//
+//  Created by BLVCK on 25/09/2017.
+//  Copyright © 2017 blvvvck production. All rights reserved.
+//
+
+import Foundation
+
+struct EducationalEstablishments {
+    var type: String
+    var name: String
+}
