+++ VKDesign(11September)/Models/Models for InformationItem/MainInfo.swift
@@ -1,0 +1,14 @@
+//
+//  MainInfo.swift
+//  VKDesign(11September)
+//
+//  Created by BLVCK on 25/09/2017.
+//  Copyright © 2017 blvvvck production. All rights reserved.
+//
+
+import Foundation
+
+struct MainInfo {
+    var title: String
+    var info: String
+}
