+++ VKDesign(11September)/Models/Models for InformationItem/MoreInfo.swift
@@ -1,0 +1,15 @@
+//
+//  MoreInfo.swift
+//  VKDesign(11September)
+//
+//  Created by BLVCK on 25/09/2017.
+//  Copyright © 2017 blvvvck production. All rights reserved.
+//
+
+import Foundation
+
+
+struct MoreInfo {
+    var title: String
+    var count: String
+}
